src/requirements.txt
asyncpg
fastapi
pydantic
SQLAlchemy
uvicorn

httpx
pytest-mock
pytest-asyncio
asgi_lifespan
mock